Begin reworking shiny hunter app Make shiny hunt processes its own class Rework hunting, plus counting logic
