+++ src/shinyhunter/embedded_app.py
